backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx==0.28.1
sqlalchemy==2.0.36
asyncpg==0.30.0
greenlet==3.1.1
